relocate organization, department module, resolver and services
