+++ apps/api/src/app.module.ts
@@ -34,6 +34,7 @@ import { validateAppConfig } from './utils/validate-app-config.util';
 import { UserModule } from './modules/user/user.module';
 import { KeycloakModule } from './modules/keycloak/keycloak.module';
 import { SettingsModule } from './modules/settings/settings.module';
+import { OrganizationModule } from './modules/organization/organization.module';
 
 @Module({
   imports: [
@@ -74,6 +75,7 @@ import { SettingsModule } from './modules/settings/settings.module';
     UserModule,
     KeycloakModule,
     SettingsModule,
+    OrganizationModule,
   ],
   controllers: [],
   providers: [
