+++ apps/api/src/modules/external/external.module.ts
@@ -1,8 +1,13 @@
 import { HttpModule } from '@nestjs/axios';
-import { Module } from '@nestjs/common';
+import { Module, forwardRef } from '@nestjs/common';
 import qs from 'qs';
 import { KeycloakModule } from '../keycloak/keycloak.module';
 import { KeycloakService } from '../keycloak/keycloak.service';
+import { DepartmentResolver } from '../organization/department/department.resolver';
+import { DepartmentService } from '../organization/department/department.service';
+import { OrganizationModule } from '../organization/organization.module';
+import { OrganizationResolver } from '../organization/organization/organization.resolver';
+import { OrganizationService } from '../organization/organization/organization.service';
 import { PrismaModule } from '../prisma/prisma.module';
 import { PrismaService } from '../prisma/prisma.service';
 import { SearchModule } from '../search/search.module';
@@ -10,14 +15,10 @@ import { SearchService } from '../search/search.service';
 import { ClassificationResolver } from './classification.resolver';
 import { ClassificationService } from './classification.service';
 import { CrmService } from './crm.service';
-import { DepartmentResolver } from './department.resolver';
-import { DepartmentService } from './department.service';
 import { LocationResolver } from './location.resolver';
 import { LocationService } from './location.service';
 import { OrgChartResolver } from './org-chart.resolver';
 import { OrgChartService } from './org-chart.service';
-import { OrganizationResolver } from './organization.resolver';
-import { OrganizationService } from './organization.service';
 import { PeoplesoftV2Service } from './peoplesoft-v2.service';
 import { PeoplesoftService } from './peoplesoft.service';
 import { PositionResolver } from './position.resolver';
@@ -29,6 +30,7 @@ import { PositionService } from './position.service';
       paramsSerializer: (params) => qs.stringify(params, { encode: false }),
     }),
     KeycloakModule,
+    forwardRef(() => OrganizationModule),
     PrismaModule,
     SearchModule,
   ],
