+++ apps/api/src/modules/external/position.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { Classification, Department } from '@prisma/client';
 import { AlexandriaError } from '../../utils/alexandria-error';
+import { DepartmentService } from '../organization/department/department.service';
+import { OrganizationService } from '../organization/organization/organization.service';
 import { ClassificationService } from './classification.service';
-import { DepartmentService } from './department.service';
 import { Employee } from './models/employee.model';
 import { FindUniquePositionArgs } from './models/find-unique-position.args';
 import { PeoplesoftPosition } from './models/peoplesoft-position.model';
 import { PositionProfile } from './models/position-profile.model';
 import { Position } from './models/position.model';
-import { OrganizationService } from './organization.service';
 import { PeoplesoftService } from './peoplesoft.service';
 
 @Injectable()
