+++ apps/api/src/modules/organization/organization/organization.service.ts
@@ -1,16 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { FindManyOrganizationArgs, FindUniqueOrganizationArgs } from '../../@generated/prisma-nestjs-graphql';
-import { PrismaService } from '../prisma/prisma.service';
+import { FindManyOrganizationArgs, FindUniqueOrganizationArgs } from '../../../@generated/prisma-nestjs-graphql';
+import { PrismaService } from '../../prisma/prisma.service';
 
 @Injectable()
 export class OrganizationService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async getOrganizations({ where, ...args }: FindManyOrganizationArgs) {
+  async getOrganizations(args?: FindManyOrganizationArgs) {
     return this.prisma.organization.findMany({
-      where: {
-        ...where,
-      },
       ...args,
       include: {
         departments: true,
