--- apps/api/src/modules/settings/department/department.resolver.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { DepartmentResolver } from './department.resolver';
-
-describe('DepartmentResolver', () => {
-  let resolver: DepartmentResolver;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [DepartmentResolver],
-    }).compile();
-
-    resolver = module.get<DepartmentResolver>(DepartmentResolver);
-  });
-
-  it('should be defined', () => {
-    expect(resolver).toBeDefined();
-  });
-});
