--- apps/api/src/modules/settings/department/department.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { DepartmentService } from './department.service';
-
-describe('DepartmentService', () => {
-  let service: DepartmentService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [DepartmentService],
-    }).compile();
-
-    service = module.get<DepartmentService>(DepartmentService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
